+++ R/dimred_dm_destiny.R
@@ -1,19 +1,19 @@
 #' @rdname dimred
 #' @export
 #'
-#' @importFrom stats as.dist
 dimred_dm_destiny <- function(x, ndim = 2) {
-  dynutils::install_packages("destiny")
+  install_packages("destiny")
+  requireNamespace("destiny")
 
-  if (dynutils::is_sparse(x)) {
+  if (is_sparse(x)) {
     x <- as.matrix(x)
   }
 
-  requireNamespace("destiny")
 
   dm <- destiny::DiffusionMap(
     data = x,
     n_eigs = ndim
   )
-  process_dimred(dm@eigenvectors[,seq_len(ndim)], rownames(x))
+
+  .process_dimred(dm@eigenvectors[,seq_len(ndim)], rownames(x))
 }