+++ R/dimred_dm_diffusionmap.R
@@ -2,15 +2,19 @@
 #' @export
 #'
 #' @importFrom stats as.dist
-dimred_dm_diffusionmap <- function(x, ndim = 2) {
-  dynutils::install_packages(dependencies = "diffusionMap", package = "dyndimred")
-
-  if (dynutils::is_sparse(x)) {
-    x <- as.matrix(x)
-  }
-
+dimred_dm_diffusionmap <- function(
+  x,
+  ndim = 2,
+  distance_metric
+) {
+  install_packages(dependencies = "diffusionMap", package = "dyndimred")
   requireNamespace("diffusionMap")
-  dist <- dynutils::correlation_distance(x)
-  space <- diffusionMap::diffuse(stats::as.dist(dist), neigen = ndim, delta = 10e-5)$X
-  process_dimred(space[,seq_len(ndim)], rownames(x))
+
+  dist <- calculate_distance(x, metric = distance_metric)
+
+  space <- diffusionMap::diffuse(as.dist(dist), neigen = ndim, delta = 10e-5)$X
+
+  .process_dimred(space[,seq_len(ndim)], rownames(x))
 }
+
+formals(dimred_dm_diffusionmap)$distance_metric <- dynutils::list_distance_metrics()