+++ R/dimred_ica.R
@@ -2,12 +2,13 @@
 #' @export
 dimred_ica <- function(x, ndim = 3) {
   dynutils::install_packages("fastICA")
+  requireNamespace("fastICA")
 
-  if (dynutils::is_sparse(x)) {
+  if (is_sparse(x)) {
     x <- as.matrix(x)
   }
 
-  requireNamespace("fastICA")
-  space <- fastICA::fastICA(t(scale(t(x))), ndim)$S
-  process_dimred(space)
+  space <- fastICA::fastICA(t(scale(t(x))), n.comp = ndim)$S
+
+  .process_dimred(space)
 }