+++ R/dimred_mds_sammon.R
@@ -1,9 +1,12 @@
 #' @rdname dimred
 #' @export
-dimred_mds_sammon <- function(x, ndim = 2) {
+dimred_mds_sammon <- function(x, ndim = 2, distance_metric) {
   dynutils::install_packages(c("MASS"), "dyndimred")
 
-  dist <- dynutils::correlation_distance(x)
-  space <- MASS::sammon(dist, k = ndim)$points
-  process_dimred(space)
+  dis <- calculate_distance(x, metric = distance_metric)
+  space <- MASS::sammon(as.dist(dis), k = ndim)$points
+
+  .process_dimred(space)
 }
+
+formals(dimred_mds_sammon)$distance_metric <- dynutils::list_distance_metrics()