+++ R/dimred_mds_smacof.R
@@ -1,9 +1,12 @@
 #' @rdname dimred
 #' @export
-dimred_mds_smacof <- function(x, ndim = 2) {
+dimred_mds_smacof <- function(x, ndim = 2, distance_metric) {
   dynutils::install_packages(c("smacof"), "dyndimred")
 
-  dist <- dynutils::correlation_distance(x)
-  space <- smacof::mds(as.dist(dist), type = "ratio", ndim = ndim)$conf
-  process_dimred(space)
+  dis <- calculate_distance(x, metric = distance_metric)
+  space <- smacof::mds(as.dist(dis), type = "ratio", ndim = ndim)$conf
+
+  .process_dimred(space)
 }
+
+formals(dimred_mds_smacof)$distance_metric <- dynutils::list_distance_metrics()