+++ R/dimred_pca.R
@@ -6,5 +6,5 @@ dimred_pca <- function(x, ndim = 2) {
 
   space <- irlba::prcomp_irlba(x, n = ndim)$x
 
-  process_dimred(space, rn = rownames(x))
+  .process_dimred(space, rn = rownames(x))
 }