+++ tests/testthat/test-dimred.R
@@ -1,9 +1,39 @@
 context("Dimred for expression")
 
-x <- seq(-1, 1, length.out = 200)
-expr <- do.call(cbind, lapply(seq_len(100), function(i) x^rbinom(1, size = 10, prob = .5) + rnorm(length(x), 0, .1)))
-rownames(expr) <- sample(paste0("row", seq_len(nrow(expr))))
-colnames(expr) <- sample(paste0("col", seq_len(ncol(expr))))
+generate_dataset <- function(type = c("splines", "polynomial"), num_samples = 400, num_genes = 500) {
+  type <- match.arg(type)
+  x <- seq(-1, 1, length.out = num_samples)
+  switch(
+    type,
+    "polynomial"={
+      y <- stats::poly(x, 2)
+      sd <- .012 * sqrt(num_genes)
+    },
+    "splines"={
+      y <- splines::ns(x, df=3)
+      sd <- .06 * sqrt(num_genes)
+    })
+  expression <- sapply(seq_len(num_genes), function(g) {
+    scale <- stats::rnorm(ncol(y), mean=0, sd=1)
+    noise <- stats::rnorm(length(x), sd=sd)
+    rowSums(sweep(y, 2, scale, "*")) + noise
+  })
+  weighted_random_sample <- function(data, weights, n){
+    key <- stats::runif(length(data)) ^ (1 / weights)
+    data[order(key, decreasing=TRUE)][seq_len(n)]
+  }
+  undetectable <- which(expression < 0)
+  undetectable <- weighted_random_sample(undetectable, -expression[undetectable], round(length(undetectable)*.5))
+
+  expression <- expression + .5
+  expression[expression < 0 | seq_along(expression) %in% undetectable] <- 0
+
+  dimnames(expression) <- list(paste0("sample", seq_len(num_samples)), paste0("feature", seq_len(num_genes)))
+  expression
+}
+
+set.seed(1)
+expr <- generate_dataset(num_samples = 400, num_genes = 200)
 
 methods <- list_dimred_methods()
 
@@ -14,6 +44,7 @@ for (method_name in names(methods)) {
 
     pdf("/dev/null")
     sink("/dev/null")
+    set.seed(1)
     out1 <- method_fun(expr)
     sink()
 
@@ -22,6 +53,7 @@ for (method_name in names(methods)) {
     expect_identical(colnames(out1), paste0("comp_", seq_len(ncol(out1))))
 
     sink("/dev/null")
+    set.seed(1)
     out2 <- method_fun(expr, ndim = 2)
     sink()
     expect_is(out2, "matrix")
@@ -30,6 +62,7 @@ for (method_name in names(methods)) {
     expect_equal(ncol(out2), 2)
 
     sink("/dev/null")
+    set.seed(1)
     ndim <- ifelse(method_name == "tsne", 3, 5)
     out3 <- dimred(expr, method = method_name, ndim = ndim)
     sink()
@@ -37,15 +70,13 @@ for (method_name in names(methods)) {
     expect_identical(rownames(out3), rownames(expr))
     expect_identical(colnames(out3), paste0("comp_", seq_len(ncol(out3))))
     expect_equal(ncol(out3), ndim)
+    dev.off()
 
     v1 <- as.vector(as.matrix(dist(out1)))
     v2 <- as.vector(as.matrix(dist(out2)))
     v3 <- as.vector(as.matrix(dist(out3)))
 
-    # even though the number of dimensions is different, one would expect the correlation between each execution to be somewhat greater than zero
-    # lle typically gives lower correlations
-    required_cor <- ifelse(method_name %in% c("lle"), 0, 0.5)
-    expect_true(all(cor(cbind(v1, v2, v3)) > required_cor))
-    dev.off()
+    expect_true(all(cor(cbind(v1, v2, v3)) > .25))
+
   })
 }