+++ inst/NEWS.md
@@ -1,3 +1,3 @@
 # dyndimred 1.0.0 (29-03-2019)
 
-* Initial release of dyndimred, includes tsne, umap, lle, (l)mds, pca and diffusion map
+* Initial release of dyndimred, includes tsne, umap, lle, pca, diffusion map, mds, landmark mds